Barnard/fyn 525 xago client

See merge request fynbos/fynbos!876

+++ go/pacioli/client/grpc.go
@@ -170,7 +170,7 @@ func (c client) GetTransfers(ctx context.Context, transferIDs []string) ([]pacio
 	for i, l := range tresp.Transfers {
 		res[i] = pacioli.Transfer{
 			ID:              l.Id,
-			LedgerID:        uint16(l.Ledger),
+			LedgerID:        l.Ledger,
 			DebitAccountID:  l.DebitAccountId,
 			CreditAccountID: l.CreditAccountId,
 			Amount:          l.Amount,
